backend/requirements.txt
torch
torchvision
Pillow                # PIL
scikit-image          #skimage
tqdm
opencv-python         # cv2
trimesh
ffmpeg
fastapi
uvicorn
matplotlib
python-multipart

#For MySQL
pydantic
sqlalchemy
mysql-connector-python
mysqlclient
python-jose
python-multipart
passlib
bcrypt